# apps/enterprise-center/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "default": {
            "description": "400 输入无效；401 身份失败；403 来源/权限拒绝；409 状态变化；429 限流；503 服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterErrorResponse"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "统一登录，返回内存 Token 和关联 LLM key；无 Cookie 操作",
        "tags": [
          "identity"
        ]
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "description": "400 输入无效；401 身份失败；403 来源/权限拒绝；409 状态变化；429 限流；503 服务不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "tokenBearer": []
          }
        ],
        "summary": "仅废弃当前 Token 和关联 key；缺失/无效/重复退出均成功；请求体为 {}",
        "tags": [
          "identity"
        ]
      }
    },
...
  },
  "info": {
    "title": "Enterprise Center API",
    "description": "企业中心已实现的 HTTP 接口；软件版本不代表 Runtime 版本。",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "tokenBearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "ct_ 内存 Token；不接受 cl_ LLM key 或 Cookie；中心重启需重新登录"
      }
    },
    "schemas": {
...
      "CenterErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "稳定原因码",
            "example": "SERVICE_UNAVAILABLE"
          },
          "message": {
            "type": "string",
            "description": "脱敏错误说明"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "description": "服务端生成的请求关联标识"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ]
      },
      "CenterErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/CenterErrorDetail"
          }
        },
        "required": [
          "error"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "3—64 位 ASCII 账号，trim 后转小写",
            "maxLength": 128
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "writeOnly": true,
            "format": "password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "IdentityUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "maximum": 2147483647
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ]
          },
          "is_super_admin": {
            "type": "boolean"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "is_super_admin",
          "enabled"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "center_id": {
            "type": "string",
            "format": "uuid"
          },
          "user": {
            "$ref": "#/components/schemas/IdentityUser"
          },
          "token": {
            "type": "string",
            "pattern": "^ct_[a-f0-9]{64}$",
            "description": "用于 Bearer 身份认证；仅内存保存，退出、撤销或中心重启后失效"
          },
          "llm_key": {
            "type": "string",
            "pattern": "^cl_[a-f0-9]{64}$",
            "description": "仅供后续 LLM 代理使用，随本次 token 一起撤销；不用于身份 API"
          }
        },
        "required": [
          "center_id",
          "user",
          "token",
          "llm_key"
        ]
      },
...
    }
  },
  "security": [
    {
      "tokenBearer": []
    }
  ]
}